refactor: group cart related services into one object with functions and fix small bugs
