+++ src/components/CartFooter.js
@@ -1,3 +1,5 @@
+import { CartContext } from "../context/CartContext.js";
+
 export const createCartFooter = () => {
     const template = document.getElementById('cart-footer-template');
     const clone = template.content.cloneNode(true);
